pyproject.toml
[project]
name = "teste-data-eng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
